Terminei as tarefas dessa semana!

+++ bd-projeto-guiado-III/gameStore/src/controller/consolesController.js
@@ -1,27 +1,69 @@
 const ConsolesModel = require("../models/consolesModel");
 
+const jwt = require ("jsonwebtoken");
+const SECRET = process.env.SECRET;
+
 const findAllConsoles = async (req, res) => {
   try {
-    const allConsoles = await ConsolesModel.find();
-    res.status(200).json(allConsoles);
+const authHeader = req.get("Authorization")
+if(!authHeader){
+  return res.status(401).send("Você esqueceu de passar as informações de autorização!")
+
+}
+
+const token = authHeader.split(" ")[1]
+
+jwt.verify(token, SECRET, async function(erro){
+  if(erro){
+    return res.status(403).send("Acesso não autorizado!")
+  }
+
+  const allConsoles = await ConsolesModel.find();
+  res.status(200).json(allConsoles);
+})
+
+    
   } catch {
     console.log(error);
     res.status(500).json({ message: error.message });
   };
 };
 
+
 const findConsoleById = async (req, res) => {
   try {
-    const findConsole = await ConsolesModel.findById(req.params.id);
-    res.status(200).json(findConsole);
+    const authHeader = req.get("Authorization")
+    if(!authHeader){
+      return res.status(401).send("Você esqueceu de passar as informações de autorização!")}
+    
+    const token = authHeader.split(" ")[1]
+    
+    jwt.verify(token, SECRET, async function(erro){
+      if(erro){
+        return res.status(403).send("Acesso não autorizado!")
+      }
+      const findConsole = await ConsolesModel.findById(req.params.id);
+      res.status(200).json(findConsole);
+    })
   } catch (error) {
     console.error(error);
     res.status(500).json({ message: error.message });
   };
-};
+}
 
 const addNewConsole = async (req, res) => {
   try {
+    const authHeader = req.get("Authorization")
+    if(!authHeader){
+      return res.status(401).send("Você esqueceu de passar as informações de autorização!")}
+    
+    const token = authHeader.split(" ")[1]
+    
+    jwt.verify(token, SECRET, async function(erro){
+      if(erro){
+        return res.status(403).send("Acesso não autorizado!")
+      }
+
     const {
       name,
       developer,
@@ -41,19 +83,31 @@ const addNewConsole = async (req, res) => {
       numberOfPlayers,
       available,
       description,
-    });
+    })
 
     const savedConsole = await newConsole.save();
 
-    res.status(201).json({ message: "New console successfully added", savedConsole });
+    res.status(201).json({ message: "New console successfully added", savedConsole })})
+
   } catch (error) {
     console.error(error);
     res.status(500).json(error.message);
-  };
-};
+  }}
+
 
 const updateConsole = async (req, res) => {
   try {
+    const authHeader = req.get("Authorization")
+    if(!authHeader){
+      return res.status(401).send("Você esqueceu de passar as informações de autorização!")}
+    
+    const token = authHeader.split(" ")[1]
+    
+    jwt.verify(token, SECRET, async function(erro){
+      if(erro){
+        return res.status(403).send("Acesso não autorizado!")
+      }
+
     const {
       name,
       developer,
@@ -73,21 +127,32 @@ const updateConsole = async (req, res) => {
       numberOfPlayers,
       available,
       description,
-    });
+    })})
 
     res.status(200).json({ message: "Console successfully updated", updateConsole });
   } catch {
     console.error(error);
     res.status(500).json({ message: error.message });
   };
-};
+}; 
 
 const deleteConsole = async (req, res) => {
   try {
+    const authHeader = req.get("Authorization")
+    if(!authHeader){
+      return res.status(401).send("Você esqueceu de passar as informações de autorização!")}
+    
+    const token = authHeader.split(" ")[1]
+    
+    jwt.verify(token, SECRET, async function(erro){
+      if(erro){
+        return res.status(403).send("Acesso não autorizado!")
+      }
+
     const { id } = req.params;
     const deleteConsole = await ConsolesModel.findByIdAndDelete(id);
     const message = `Console with id ${deleteConsole.name} was successfully deleted`;
-    res.status(200).json({ message });
+    res.status(200).json({ message })})
   } catch (error){
     console.error(error);
     res.status(500).json({ message: error.message });
@@ -99,5 +164,10 @@ module.exports = {
   findConsoleById,
   addNewConsole,
   updateConsole,
-  deleteConsole,
-};
+  deleteConsole
+}
+
+/*O código Update by Id aconteceu o seguinte: O postman me devolvia 200 ok e a mensagem de que 
+o console havia sido atualizado com sucesso, mas no vs code o codigo quebrava, quando limpo o terminal
+e dou o npm start no get all consoles ele nao aparece atualizado, os demais codigos não apresentaram
+problemas  =)  */
